Added the addDays method to the readme
I was confused about whether it was possible or not to add and subtract days since it seemed to be deliberately left out, espescially I read the vanillaJS docs right before and the way of setting and getting the days is a bit "unique". Long story short: Added the addDays method to the readme and made the comments consistent.